Create separate class for ShellValue
